refactor(app-server): extract token usage replay helpers
Move restored token usage replay logic into a focused
`token_usage_replay` submodule. The message processor keeps ownership
of resume and fork lifecycle ordering while the helper module owns notification construction and rollout-based turn attribution.
