+++ codex-rs/app-server/src/codex_message_processor.rs
@@ -136,7 +136,6 @@ use codex_app_server_protocol::ThreadDecrementElicitationParams;
 use codex_app_server_protocol::ThreadDecrementElicitationResponse;
 use codex_app_server_protocol::ThreadForkParams;
 use codex_app_server_protocol::ThreadForkResponse;
-use codex_app_server_protocol::ThreadHistoryBuilder;
 use codex_app_server_protocol::ThreadIncrementElicitationParams;
 use codex_app_server_protocol::ThreadIncrementElicitationResponse;
 use codex_app_server_protocol::ThreadInjectItemsParams;
@@ -178,8 +177,6 @@ use codex_app_server_protocol::ThreadStartParams;
 use codex_app_server_protocol::ThreadStartResponse;
 use codex_app_server_protocol::ThreadStartedNotification;
 use codex_app_server_protocol::ThreadStatus;
-use codex_app_server_protocol::ThreadTokenUsage;
-use codex_app_server_protocol::ThreadTokenUsageUpdatedNotification;
 use codex_app_server_protocol::ThreadUnarchiveParams;
 use codex_app_server_protocol::ThreadUnarchiveResponse;
 use codex_app_server_protocol::ThreadUnarchivedNotification;
@@ -363,12 +360,17 @@ use codex_app_server_protocol::ServerRequest;
 mod apps_list_helpers;
 mod plugin_app_helpers;
 mod plugin_mcp_oauth;
+mod token_usage_replay;
 
 use crate::filters::compute_source_filters;
 use crate::filters::source_kind_matches;
 use crate::thread_state::ThreadListenerCommand;
 use crate::thread_state::ThreadState;
 use crate::thread_state::ThreadStateManager;
+use token_usage_replay::latest_token_usage_turn_id_for_thread_path;
+use token_usage_replay::latest_token_usage_turn_id_from_rollout_items;
+use token_usage_replay::latest_token_usage_turn_id_from_rollout_path;
+use token_usage_replay::send_thread_token_usage_update_to_connection;
 
 const THREAD_LIST_DEFAULT_LIMIT: usize = 25;
 const THREAD_LIST_MAX_LIMIT: usize = 100;
@@ -8601,107 +8603,6 @@ async fn handle_pending_thread_resume_request(
     outgoing
         .replay_requests_to_connection_for_thread(connection_id, conversation_id)
         .await;
-}
-
-/// Sends a restored token usage update to the connection that attached to a thread.
-///
-/// This is lifecycle replay rather than a model event: the rollout already contains
-/// the original `TokenCount`, and emitting through `send_event` here would duplicate
-/// persisted usage records. Keeping this helper connection-scoped also avoids
-/// surprising other subscribers with a historical usage update while they may be
-/// rendering live turn events.
-async fn send_thread_token_usage_update_to_connection(
-    outgoing: &Arc<OutgoingMessageSender>,
-    connection_id: ConnectionId,
-    thread_id: ThreadId,
-    thread: &Thread,
-    conversation: &CodexThread,
-    token_usage_turn_id: Option<String>,
-) {
-    let Some(info) = conversation.token_usage_info().await else {
-        return;
-    };
-    let notification = ThreadTokenUsageUpdatedNotification {
-        thread_id: thread_id.to_string(),
-        turn_id: token_usage_turn_id.unwrap_or_else(|| latest_token_usage_turn_id(thread)),
-        token_usage: ThreadTokenUsage::from(info),
-    };
-    outgoing
-        .send_server_notification_to_connections(
-            &[connection_id],
-            ServerNotification::ThreadTokenUsageUpdated(notification),
-        )
-        .await;
-}
-
-async fn latest_token_usage_turn_id_for_thread_path(thread: &Thread) -> Option<String> {
-    let rollout_path = thread.path.as_deref()?;
-    latest_token_usage_turn_id_from_rollout_path(rollout_path, thread).await
-}
-
-async fn latest_token_usage_turn_id_from_rollout_path(
-    rollout_path: &Path,
-    thread: &Thread,
-) -> Option<String> {
-    let rollout_items = read_rollout_items_from_rollout(rollout_path).await.ok()?;
-    latest_token_usage_turn_id_from_rollout_items(&rollout_items, thread)
-}
-
-struct TokenUsageTurnOwner {
-    id: String,
-    position: Option<usize>,
-}
-
-fn latest_token_usage_turn_id_from_rollout_items(
-    rollout_items: &[RolloutItem],
-    thread: &Thread,
-) -> Option<String> {
-    let owner = latest_token_usage_turn_owner_from_rollout_items(rollout_items)?;
-    if thread.turns.iter().any(|turn| turn.id == owner.id) {
-        return Some(owner.id);
-    }
-    owner
-        .position
-        .and_then(|position| thread.turns.get(position))
-        .map(|turn| turn.id.clone())
-}
-
-fn latest_token_usage_turn_owner_from_rollout_items(
-    rollout_items: &[RolloutItem],
-) -> Option<TokenUsageTurnOwner> {
-    let mut builder = ThreadHistoryBuilder::new();
-    let mut token_usage_turn_owner = None;
-
-    for item in rollout_items {
-        if matches!(item, RolloutItem::EventMsg(EventMsg::TokenCount(_))) {
-            token_usage_turn_owner =
-                builder
-                    .active_turn_snapshot()
-                    .map(|turn| TokenUsageTurnOwner {
-                        id: turn.id,
-                        position: builder.active_turn_position(),
-                    });
-        }
-        builder.handle_rollout_item(item);
-    }
-
-    token_usage_turn_owner
-}
-
-/// Chooses a fallback turn id that should own a replayed token usage update.
-///
-/// Normal replay derives the owner from the rollout position of the latest
-/// `TokenCount` event. This fallback only preserves a stable wire shape for
-/// unusual histories where that rollout information cannot be read.
-fn latest_token_usage_turn_id(thread: &Thread) -> String {
-    thread
-        .turns
-        .iter()
-        .rev()
-        .find(|turn| matches!(turn.status, TurnStatus::Completed | TurnStatus::Failed))
-        .or_else(|| thread.turns.last())
-        .map(|turn| turn.id.clone())
-        .unwrap_or_default()
 }
 
 enum ThreadTurnSource<'a> {
